constructor for arduino changed

+++ src/FlashMemory.h
@@ -39,7 +39,7 @@ class Flash
 {
   public:
     Flash(uint8_t cs);
-    Flash(uint8_t mosi, uint8_t miso, uint8_t sck, uint8_t cs);
+    Flash(uint32_t mosi, uint32_t miso, uint32_t sck, uint32_t cs);
     void begin();
     void read(uint32_t addr, uint8_t *buf, uint16_t len);
     void write(uint32_t addr, uint8_t *buf, uint16_t len);
